+++ vm/src/types/relocatable.rs
@@ -7,6 +7,7 @@ use crate::Felt252;
 use crate::{
     relocatable, types::errors::math_errors::MathError, vm::errors::memory_errors::MemoryError,
 };
+use num_bigint::{BigInt, BigUint};
 use num_traits::ToPrimitive;
 use serde::{Deserialize, Serialize};
 
@@ -58,12 +59,6 @@ impl From<(isize, usize)> for MaybeRelocatable {
     }
 }
 
-impl From<usize> for MaybeRelocatable {
-    fn from(num: usize) -> Self {
-        MaybeRelocatable::Int(Felt252::from(num))
-    }
-}
-
 impl From<Felt252> for MaybeRelocatable {
     fn from(num: Felt252) -> Self {
         MaybeRelocatable::Int(num)
@@ -93,6 +88,25 @@ impl From<Relocatable> for MaybeRelocatable {
         MaybeRelocatable::RelocatableValue(rel)
     }
 }
+
+// Implement primitive and big-int (owned + reference) conversions by first converting to Felt252,
+// then wrapping as MaybeRelocatable::Int.
+macro_rules! impl_from_for_maybe_relocatable {
+    ($($t:ty),* $(,)?) => {
+        $(
+            impl From<$t> for MaybeRelocatable {
+                fn from(num: $t) -> Self {
+                    MaybeRelocatable::Int(Felt252::from(num))
+                }
+            }
+        )*
+    };
+}
+
+impl_from_for_maybe_relocatable!(
+    u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128, isize, BigUint, BigInt, &BigUint,
+    &BigInt
+);
 
 impl Display for MaybeRelocatable {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {